Add files via upload
https://mega.nz/file/db5BFLhC#xrOfc-PXFmKnrOvmTBBFxZjmCThO4dnsbECeT1XlhPY

I had lots of thermal run issues and after viewing the video provided all the issue went away.
also, i added a mega link with the files that worked for me. Arduino IDE wasn't working for me 
so i also added the ANET Board drivers and IDE that works to downgrade and upgrade your 
board.

Took me a while to find all this and lots of frustration, i hope it works for you as it did for me.

I will fix the files later, it is not in order.